# T::Utils.coerce: Push special-case module->T::Types::Base logic behind cache