--- a/gems/sorbet-runtime/lib/types/types/union.rb
+++ b/gems/sorbet-runtime/lib/types/types/union.rb
@@ -8,12 +8,12 @@ module T::Types
 
     def initialize(types)
       @types = types.flat_map do |type|
-        type = T::Utils.resolve_alias(type)
+        type = T::Utils.coerce(type)
         if type.is_a?(Union)
           # Simplify nested unions (mostly so `name` returns a nicer value)
           type.types
         else
-          T::Utils.coerce(type)
+          type
         end
       end.uniq
     end