--- a/gems/sorbet-runtime/lib/types/utils.rb
+++ b/gems/sorbet-runtime/lib/types/utils.rb
@@ -8,18 +8,6 @@ module T::Utils
       val.aliased_type
     elsif val.is_a?(T::Types::Base)
       val
-    elsif val == ::Array
-      T::Array[T.untyped]
-    elsif val == ::Set
-      T::Set[T.untyped]
-    elsif val == ::Hash
-      T::Hash[T.untyped, T.untyped]
-    elsif val == ::Enumerable
-      T::Enumerable[T.untyped]
-    elsif val == ::Enumerator
-      T::Enumerator[T.untyped]
-    elsif val == ::Range
-      T::Range[T.untyped]
     elsif val.is_a?(Module)
       T::Types::Simple::Private::Pool.type_for_module(val)
     elsif val.is_a?(::Array)