First commit: When coercing a module to a type, we do a bunch of checks for specific special classes like Array before falling back to the generic `T::Types::Simple.new`. But these are edge cases that don't occur often, plus the same caching mechanism we use for Simple can work here, so we can speed up `T.let` and friends in the common case by pushing these checks into `type_for_module` and caching the result.

Second commit: Lower impact and only loosely related, but since `T::Utils.coerce` already resolves aliases there's no need to invoke `resolve_alias` separately when initializing a `T::Types::Union`.

### Motivation
Before (relevant excerpt):
```
 sorbet-runtime git:(master) ✗ bundle exec rake bench:typecheck
...
T.let(..., Integer): 431.782 ns
T.let(..., T.nilable(Integer)): 1.129 μs
T.let(..., Example): 425.731 ns
T.let(..., T.nilable(Example)): 1.119 μs
```

After:
```
sorbet-runtime git:(djudd/coerce-opts) ✗ bundle exec rake bench:typecheck
...
T.let(..., Integer): 304.652 ns
T.let(..., T.nilable(Integer)): 1.001 μs
T.let(..., Example): 310.781 ns
T.let(..., T.nilable(Example)): 1.028 μs
```

### Test plan
Existing tests